backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg[binary]>=3.1.18

# Redis & Celery
celery>=5.3.6
redis>=5.0.1

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# File handling
openpyxl>=3.1.2
pandas>=2.2.0

# PDF generation
reportlab>=4.0.8

# WebSocket
websockets>=12.0

# Utils
python-dotenv>=1.0.0
httpx>=0.26.0
